refactor component filenames and sdkinstance type keys
