+++ packages/paypal-js/types/v6/components/paypal-legacy-billing-agreemens.ts
@@ -3,7 +3,7 @@ import {
     OnCancelData,
     OnErrorData,
     OnApproveDataOneTimePayments,
-} from "./payments";
+} from "./paypal-payments";
 
 export type BillingSessionInputs = {
     onApprove?: (data: OnApproveDataOneTimePayments) => Promise<void>;
