+++ packages/paypal-js/types/v6/components/venmo-payments.ts
@@ -1,0 +1,6 @@
+import { PayPalOneTimePaymentPaymentSessionInputs } from "./paypal-payments";
+
+export type VenmoPaymentSessionInputs = Omit<
+    PayPalOneTimePaymentPaymentSessionInputs,
+    "onShippingAddressChange" | "onShippingOptionsChange"
+>;
